hello check it

+++ src/main/java/com/awesome/repository/StoreRepository.java
@@ -3,7 +3,7 @@ package com.awesome.repository;
 import com.awesome.domain.Stores;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StoreRepository extends CrudRepository<Stores,Integer>{
+public interface StoreRepository extends CrudRepository<Stores,Long>{
 
 }
